# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplikasi antrean",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login User",
                "description": "Endpoint untuk melakukan autentikasi dan mendapatkan token akses.",
                "operationId": "6722e46f789eb246c9c35010d18e84c9",
                "requestBody": {
                    "description": "Data login yang diperlukan.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Alamat email pengguna.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@admin.com"
                                    },
                                    "password": {
                                        "description": "Password pengguna.",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil login dan mendapatkan token akses.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status response.",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "description": "Token akses yang digunakan untuk otentikasi.",
                                                    "type": "string",
                                                    "example": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImp0aSI6IjlmNjE5NjM3YzAxZ..."
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "description": "ID pengguna.",
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "description": "Nama pengguna.",
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "email": {
                                                            "description": "Alamat email pengguna.",
                                                            "type": "string",
                                                            "format": "email",
                                                            "example": "user@example.com"
                                                        },
                                                        "assignment": {
                                                            "description": "ID tugas yang diberikan kepada pengguna.",
                                                            "type": "integer",
                                                            "example": 123
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Gagal autentikasi, email atau password salah.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status response.",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Pesan error.",
                                            "type": "string",
                                            "example": "Email atau password salah"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register User",
                "description": "Endpoint untuk mendaftarkan pengguna baru.",
                "operationId": "71613a537edb63807ecc9773ad6bd3af",
                "requestBody": {
                    "description": "Data yang diperlukan untuk mendaftarkan pengguna baru.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nama pengguna.",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Alamat email pengguna.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "description": "Password pengguna (minimal 6 karakter).",
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pengguna berhasil terdaftar.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Pesan sukses.",
                                            "type": "string",
                                            "example": "User registered successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Gagal mendaftarkan pengguna karena validasi gagal.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Pesan error.",
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "description": "Daftar error validasi untuk nama.",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The name field is required."
                                                    ]
                                                },
                                                "email": {
                                                    "description": "Daftar error validasi untuk email.",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The email field is required.",
                                                        "The email has already been taken."
                                                    ]
                                                },
                                                "password": {
                                                    "description": "Daftar error validasi untuk password.",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The password field is required.",
                                                        "The password must be at least 6 characters."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout User",
                "description": "Endpoint untuk logout pengguna dan mencabut token akses saat ini.",
                "operationId": "a9c63ed7d52168af81d5532dcce9eff8",
                "responses": {
                    "200": {
                        "description": "Berhasil logout pengguna.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status response.",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Pesan sukses.",
                                            "type": "string",
                                            "example": "Berhasil logout"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Gagal autentikasi, token tidak valid atau tidak ada token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Pesan error.",
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}